+++ Examples/ProductSample/Domain/UseCases/SignInUseCase.swift
@@ -7,10 +7,6 @@
 
 import Foundation
 
-struct Credentials {
-  let email, password: String
-}
-
 protocol SignInUseCase: UseCase<Credentials, Task<Void, Error>> {}
 
 struct SignInUseCaseImpl: SignInUseCase {
@@ -19,18 +15,6 @@ struct SignInUseCaseImpl: SignInUseCase {
   func execute(input: Credentials) -> Task<(), Error> {
     Task {
       try await repository.signIn(email: input.email, password: input.password)
-    }
-  }
-}
-
-protocol SignUpUseCase: UseCase<Credentials, Task<SignUpResult, Error>> {}
-
-struct SignUpUseCaseImpl: SignUpUseCase {
-  let repository: AuthenticationRepository
-
-  func execute(input: Credentials) -> Task<SignUpResult, Error> {
-    Task {
-      try await repository.signUp(email: input.email, password: input.password)
     }
   }
 }
